Refactor file block to contain dropzone block
