+++ projects/packages/forms/src/blocks/contact-form/components/jetpack-field-file/index.js
@@ -8,89 +8,6 @@ import JetpackFieldControls from '../jetpack-field-controls';
 import JetpackFieldLabel from '../jetpack-field-label';
 import { useJetpackFieldStyles } from '../use-jetpack-field-styles';
 
-import './editor.css';
-
-const DEFAULT_ICON = `${ window?.jpFormsBlocks?.defaults?.assetsUrl }/images/upload-icon.svg`;
-
-const BLOCKS_TEMPLATE = [
-	[
-		'core/group',
-		{
-			style: {
-				spacing: {
-					padding: {
-						top: '48px',
-						bottom: '48px',
-						left: '48px',
-						right: '48px',
-					},
-					margin: {
-						top: '8px',
-						bottom: '8px',
-					},
-				},
-				border: {
-					style: 'dashed',
-					width: '1px',
-					color: 'rgba(125,125,125,0.3)',
-				},
-			},
-		},
-		[
-			[
-				'core/image',
-				{
-					url: DEFAULT_ICON,
-					width: 24,
-					height: 24,
-					align: 'center',
-					className: 'is-style-default',
-					style: {
-						spacing: {
-							margin: {
-								bottom: '20px',
-							},
-						},
-					},
-				},
-			],
-			[
-				'core/paragraph',
-				{
-					align: 'center',
-					content: __(
-						'<strong><a href="#">Select a file</a></strong> or drag and drop your file here',
-						'jetpack-forms'
-					),
-					style: {
-						spacing: {
-							padding: {
-								top: '8px',
-								bottom: '8px',
-							},
-						},
-						typography: {
-							fontSize: '16px',
-						},
-					},
-				},
-			],
-			[
-				'core/paragraph',
-				{
-					align: 'center',
-					content: __( 'JPEG, PNG, PDF, and MP4 formats', 'jetpack-forms' ),
-					style: {
-						typography: {
-							fontSize: '14px',
-						},
-					},
-				},
-			],
-		],
-	],
-];
-
 const JetpackFieldFile = props => {
 	const { attributes, clientId, isSelected, setAttributes } = props;
 	const { id, label, required, requiredText, width } = attributes;
@@ -113,12 +30,10 @@ const JetpackFieldFile = props => {
 					requiredText={ requiredText }
 					setAttributes={ setAttributes }
 				/>
-				<div className="jetpack-form-file-field__dropzone">
-					<div className="jetpack-form-file-field__dropzone-inner">
-						<input type="file" style={ { display: 'none' } } aria-hidden="true" />
-						<InnerBlocks template={ BLOCKS_TEMPLATE } templateLock={ false } />
-					</div>
-				</div>
+				<InnerBlocks
+					template={ [ [ 'jetpack/field-file-dropzone', {}, [] ] ] }
+					templateLock="all"
+				/>
 			</div>
 
 			<JetpackFieldControls
